Split out REPL for use in Dialog
